icewind: final name for caom2-meta-sync
