--- darkspire/Dockerfile
@@ -1,5 +1,0 @@
-FROM images.opencadc.org/library/cadc-java:1
-
-ADD build/distributions/darkspire.tar /
-
-CMD ["/darkspire/bin/darkspire"]
