+++ icewind/Dockerfile
@@ -1,0 +1,5 @@
+FROM images.opencadc.org/library/cadc-java:1
+
+ADD build/distributions/icewind.tar /
+
+CMD ["/icewind/bin/icewind"]
