+++ icewind/build.gradle
@@ -14,7 +14,7 @@ sourceCompatibility = 1.8
 
 group = 'org.opencadc'
 
-mainClassName = 'org.opencadc.darkspire.Main'
+mainClassName = 'org.opencadc.icewind.Main'
 
 dependencies {
     implementation 'org.opencadc:cadc-util:[1.6,2.0)'
