+++ icewind/src/main/java/org/opencadc/icewind/ObservationHarvester.java
@@ -67,7 +67,7 @@
  ************************************************************************
  */
 
-package org.opencadc.darkspire;
+package org.opencadc.icewind;
 
 import ca.nrc.cadc.caom2.Artifact;
 import ca.nrc.cadc.caom2.Observation;
